Refactor(SCIMMY.Messages.BulkOp): end renaming BulkOp to BulkRequest
